Refactor spi send/read for DESFire methods
Send a command and read the result is now a single method
called desfire_sendAndRead which takes the command length and
the data length and calculates the correct size of the
data to fetch. The buffer used is the already used pn532_packetbuffer.

Also introduced is a method to fetch the error from the DESFire packet.
This is not the error code from the PN532 but the error sent by
the DESFire chip, so the communication went right but something
was wrong with the semantics and the card complained.